+++ watchapedia/app/auth/utils.py
@@ -1,7 +1,7 @@
 from datetime import datetime, timedelta
-from app.auth.settings import JWT_SETTINGS
+from watchapedia.app.auth.settings import JWT_SETTINGS
 from jose.exceptions import JWTError, ExpiredSignatureError
-from common.errors import ExpiredSignatureError, InvalidTokenError
+from watchapedia.common.errors import ExpiredSignatureError, InvalidTokenError
 from passlib.context import CryptContext
 import jwt
