+++ watchapedia/app/user/dto/requests.py
@@ -1,7 +1,7 @@
 from pydantic import BaseModel
 from typing import Annotated
 from pydantic.functional_validators import AfterValidator
-from app.auth.dto.requests import validate_username, validate_password
+from watchapedia.app.auth.dto.requests import validate_username, validate_password
 
 class UserUpdateRequest(BaseModel):
     username: Annotated[str | None, AfterValidator(validate_username)] = None
