+++ watchapedia/database/__init__.py
@@ -1,0 +1,1 @@
+import watchapedia.app.user.models
