+++ watchapedia/settings.py
@@ -2,8 +2,8 @@ import os
 
 from pydantic_settings import BaseSettings
 
-ENV = os.getenv("ENV", "local")
-assert ENV in ("local")
+ENV = os.getenv("ENV", "prod") # 배포 시에는(merge) prod로 바꿔주세요
+assert ENV in ("local", "prod")
 
 
 class Settings(BaseSettings):
